feat: edit changeRepresentativeCard
