+++ packages/web/src/features/manage-club/frames/_atomic/RepresentativeLoadFrame.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 
 import AsyncBoundary from "@sparcs-clubs/web/common/components/AsyncBoundary";
-import ChangeRepresentativeCardV1 from "@sparcs-clubs/web/features/manage-club/components/ChangeRepresentativeCardV1";
+import ChangeRepresentativeCard from "@sparcs-clubs/web/features/manage-club/components/ChangeRepresentativeCard";
 import { useGetClubDelegate } from "@sparcs-clubs/web/features/manage-club/services/getClubDelegate";
 import { useGetDelegateCandidates } from "@sparcs-clubs/web/features/manage-club/services/getDelegateCandidates";
 
@@ -21,12 +21,11 @@ const RepresentativeLoadFrame: React.FC<{
   } = useGetDelegateCandidates({ clubId, delegateEnumId: 1 });
 
   return (
-    // TODO: ChangeRepresentativeCard로 변경
     <AsyncBoundary
       isLoading={isLoading || clubMembersIsLoading}
       isError={isError || clubMembersIsError}
     >
-      <ChangeRepresentativeCardV1
+      <ChangeRepresentativeCard
         clubId={clubId}
         delegatesNow={delegatesNow || { delegates: [] }}
         clubMembers={clubMembers || { students: [] }}
