+++ packages/web/src/features/my/components/ChangeRepresentativeModalContent.tsx
@@ -18,6 +18,7 @@ interface ChangeRepresentativeModalContentProps {
   phonePlaceholder?: string;
   onClose: () => void;
   refetch: () => void;
+  requestId: number;
 }
 
 const ButtonWrapper = styled.div`
@@ -36,6 +37,7 @@ const ChangeRepresentativeModalContent: React.FC<
   phonePlaceholder = "010-XXXX-XXXX",
   onClose,
   refetch,
+  requestId,
 }) => {
   const [errorPhone, setErrorPhone] = useState<boolean>(false);
   const [phone, setPhone] = useState<string>("");
@@ -43,7 +45,7 @@ const ChangeRepresentativeModalContent: React.FC<
   // TODO: clb013 014 수정되면 반영
   const onConfirm = () => {
     patchMyDelegateRequest(
-      { requestId: 1 },
+      { requestId },
       {
         phoneNumber: phone,
         clubDelegateChangeRequestStatusEnum:
@@ -56,7 +58,7 @@ const ChangeRepresentativeModalContent: React.FC<
 
   const onReject = () => {
     patchMyDelegateRequest(
-      { requestId: 1 },
+      { requestId },
       {
         clubDelegateChangeRequestStatusEnum:
           ClubDelegateChangeRequestStatusEnum.Rejected,
